+++ CHANGELOG.md
@@ -34,6 +34,8 @@
     * Introduce background operations for _local_ reactivate/deactivate (\#2033).
     * Fix escaping of newlines within f-strings, in logs (\#2028).
     * Improve handling of task groups created before 2.9.0 (\#2050).
+* Runner:
+    * Improve error handling in `_zip_folder_to_file_and_remove` (\#2057).
 * SSH internals:
     * Add `FractalSSH.remote_exists` method (\#2008).
     * Try re-opening the connection in `FractalSSH.check_connection` when an error occurs (\#2035).